backend/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
APScheduler==3.9.1.post1
apturl==0.5.2
asgiref==3.5.0
attrs==21.2.0
autopep8==2.0.0
Babel==2.11.0
beautifulsoup4==4.10.0
bitarray==2.6.1
blinker==1.4
cached-property==1.5.2
catfish==4.16.3
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
django-cors-headers==3.13.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
esbonio==0.14.1
gpg===1.16.0-unknown
greenlet==2.0.1
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
ImageHash==4.3.1
imagesize==1.4.1
importlib-metadata==4.6.4
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightdm-gtk-greeter-settings==1.2.2
lxml==4.8.0
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
menulibre==2.2.2
more-itertools==8.10.0
mugshot==0.4.3
mysql==0.0.3
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
packaging==21.3
pexpect==4.8.0
Pillow==9.0.1
ply==3.11
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.9.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.2
pygls==0.13.0
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyspellchecker==0.7.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scipy==1.9.3
scour==0.38.2
SecretStorage==3.3.1
sgt-launcher==0.2.7
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.45
sqlparse==0.4.2
systemd-python==234
terminator==2.1.1
texttable==1.6.4
TLPUI==1.5.0
tomli==2.0.1
ttystatus==0.38
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
xcffib==0.11.1
xdg==5
xkit==0.0.0
zipp==1.0.0
